modify: 사전렌더링 막음으로써 server에서 모각코 디테일 정보 요청 막음(이해가 안됨..)
