+++ src/app/_components/Provider.tsx
@@ -12,6 +12,7 @@ const Provider = ({
   const queryClient = new QueryClient({
     defaultOptions: {
       queries: {
+        staleTime: 1000 * 60,
         refetchOnWindowFocus: false,
         retry: false,
       },
